Add back Profiler tool window
IDEA merge has brought behavior changes in ToolWindowFactory that
only shows a tool window if isAvailable=true.

However with this fix, we won't be able to hide the profiler tool
window on start-up: if it was open before user closed the project
it will be reopened the next time user opens the project, just
like any other tool window.

Bug: 155128891
Test: manual
Change-Id: Id9dd18c48a50f8433a803fc52d999d54a6434cc6