Fix race condition on pooled image check
This CL does not fix the race 100% but should address the problem in
most cases.
Before this CL, we would check if the RenderResult had a valid image and
then operate on it. The image could be disposed in the time. Now, we get
the image and we check if it's valid.

Fixes: 155190702
Bug: 155258314
Test: UI tests now pass
Change-Id: I726bca726651938abc437167d085e788591675c1