Simplify AtraceFrame as we only add one Slice per Frame
A lot of AtraceFrame class logic was dynamic, based on the fact
that a bunch of properties could change when addSlice() was called.

But right now we only add one slice per frame, so we can simplify a
lot of the logic inside the class. The only mutability inside the class
right now is when we set an associated frame.

Bug: cleanup
Test: existing
Change-Id: I92149f19326de22509de5b1ce0a50d1047fb97cc