Configure "Generate *.iml files" in all cases
There are several code execution paths that create/configure projects
in Android Studio: (1) Opening a project; (2) create a new project;
(3) importing a project from an old format and (4) setting up a project
for tests. Some of them do not use GradleProjectImporter.createProject
and other do not use GradleProjectImporter.importProjectNoSync. This is
for valid reasons. However there are some settings including
`setStoreProjectFilesExternally(false)` that we need to
configure in both cases. This change extracts them into a separate
method which will be called in all the cases.

Bug: n/a
Test: existing
Change-Id: If1af6cf350841f51f98a32908160fa93287dc996