Fix Placeholder tests
The MergingUpdateQueue has changed the default behaviour after the latest
IntelliJ merge. By default, when running unit tests, all queues would be
automatically setup in passthrough mode but that is not the case
anymore.
This CL allows the SyncLayoutlibSceneManager to do additional setup in
the queue so it can enable the passthrough mode.

Fixes: 155274325
Test: Re-enabled tests pass.
Change-Id: I2f028a1e9e46fc41681bc10d41d2e26f9f848388