Update view scrolling inside a render action
When a user scrolls a ScrollView, we need to update the view position and also render
again. That can lead to problems if the scrolling tries to update a view that is
currently being modified by the rendering. To avoid such issues, we run the scrolling
update inside a render action.

Bug: 155339633
Test: Could not manually reproduce the issue with this change
Change-Id: Ibb2f3c08b9e014bec156e245e325129fc176c5c0