[BA] Add empty states to Warnings and Tasks
It is possible that there are no warnings generated for the build.
Thus we need to show it as a special state to explain user it is
valid and not broken.
It is also possible that there are no tasks to show for the build.
It is possible if build is incremental and all tasks are up to date.

Bug: 155463804
Test: Added unit test for views
Change-Id: I7cd043e69382b6cbcbf9ad48b08f3e614976b550