Fix ArchiveToGradleModuleModelTest on Windows
After Change I55b682ca8, the Windows native file system is enabled
in Bazel tests (just as it is in production). Unfortunately this
caused ArchiveToGradleModuleModelTest to fail on Windows because
BasicJarHandler caches jar file handles for a couple seconds, thereby
preventing the File.move() that occurs in this test.

We need to fix this (possibly in the platform), but until then we can
unblock Change I55b682ca8 by calling
JarFileSystemImpl.cleanupForNextTest before File.move().

Bug: 155676674
Test: ArchiveToGradleModuleModelTest
Change-Id: I67c1c87632bbbaedbe92dec24d9fde4bae6c7ae2