Click referenced id in component tree animates the view
As titled. It introduces a highlighted selection to the model. It is added
to distinguish it from the normal drawing state.

There's one minor bug where if referenced item is dnd'd, the selection is
lost in the component tree. This is an innate bug with component and has
nothing to do with this CL, but will follow up later.

Bug: 154027782
Test: Manually tested.
1) clicking individual referenced items draw animated frames
2) clicking multiple referenced items
3) clicking referenced items and non-referenced items
4) clicking referenced item and the original item (original selection
takes higher priority)

Change-Id: I2670cc42b3c6d490a62ea06dbc2b9c8a70c929cb