Fix allocation tracking line when native memory is enabled
The issue here was the "else if" meaning that native allocation tracking would
override live allocation tracking in all cases when native allocation tracking
was enabled. When the else if was removed the native allocation tracking was
rendered between the background layer and the live allocation tracking layer
to compensate for this the rendered was added after to be drawn later than the
live allocation tracking renderer.

Test: manual, need to expose some internal state to properly test
Fixes: 155903556
Change-Id: I18d88d21060137f6e1d2e1ee365683e65e2bd712