[Dagger] Changes custom name for consumers and component's method in RelatedItemsPopup.
Consumers are parameters or properties.
In most cases consumers of one type has the same name. It leads to showing few identical names in a RelatedItemsPopup.
CL changes custom name to show function or class name to make navigation information in RelatedItemsPopup more useful.

The same for Category "Expose in components". CL changes components method's name to component name.

This Cl changes only string in RelatedItemsPopup.
Target elements are still parameters or properties for consumers and component's method for "Expose in components.

Fixes: 155961111
Test: DaggerRelatedItemLineMarkerProviderTest.kt
Change-Id: I34c3c0f16ac72141fa322bf06c435c46db425c8a