Removed no longer needed LightModelClassListProvider
It was introduced at ag/10848294 and I guess it works now is because we
have removed the module level ModelFileModificationTracker which has
reference to the myLightModelClassMap cache.

Also decided to not delete MlModuleService for now and leave it
serve for the actual light class generation since it requires module
info.

Bug: 155121511
Test: code cleanup
Change-Id: I7c3b40714e85477d75772741de623dfc059dd325