Keep our default language level to 1.6 in tests
After commit b9eddfc9fd52, the default language level in
JavaProjectData is no longer initialized to 1.6 (the default),
but the highest. As a result, we end up running tests with 13 as
the language target, even though we obtain the correct level from
Gradle sync. This change just makes the default setting we relied
on explicit.

See: https://youtrack.jetbrains.com/issue/IDEA-180986
Fixes: 155929878
Change-Id: I7aa7d750b917b6c26ea09fc1d47bd2352b736dee