Reset cursor when `RangeSelectionComponent` loses focus
This CL fixes a bug described in comment #2 in b/154264530
where if the user clicks the heap dump
in the timeline and opens a new stage, the cursor remains
a range selection one because `mouseExited` is never triggered.

The fix is by listening to hierarchy change in the
`RangeSelectionComponent`. Other obvious solutions that were
tried but didn't work (or only partially worked) are listening to
`ComponentListener.componentHidden()` (it's never called)
and `FocusListener.focusLost()` (component doesn't always
have focus in the first place).

Bug: 154264530
Test: add new
Change-Id: I6f73a42981d827f35641ff4a4e6955886a68c717