Fix hang in SdkComponentSourceTest
It seems like hangs were happening because tests were needlessly being
run on the UI thread (due to extending AndroidTestCase or
UsefulTestCase), and invoking dispatchAllInvocationEvents.
Switched to all JUnit 4, causing them not to run on the UI thread and
not need to dispatchAllInvocationEvents.
This seems to have had the side effect of making these tests much faster
as well.

Fixes: 156107369
Test: this is it
Change-Id: Iff62822c4de93eb3bf124248408f2b1aba3f1c71