Better error handling
Improved the heuristic to detect errors coming from the
ComposeViewAdapter. In some cases, the ComposeViewAdapter will throw in
the onLayout and the old detection code would not flag it as an error.
With this change, we detect those cases better.

Also moved some internal utility methods to a separate file separated
from Preview.kt.

Bug: 154609017
Test: Manually checked
Change-Id: I07595e129b792766695ad5703dabdc3a766a749e