[Dsl] represent the implicit debug signingConfig
Create the implicitly-created "debug" signing configuration in its
place in the Dsl element hierarchy when the user requests the
signingConfigs model.  This causes some knock-on effects in tests,
where the implicit signingConfig was implicitly assumed not to be
known to the parsed model.

Include tests of some scenarios of renaming signingConfigs to or from
"debug" to check that the backend writers do the right thing.

Bug: 155497681
Test: existing tests updated, fixed
Test: new SigningConfigModel tests
Change-Id: Ie770db7db6a363958fdaa5951955f10cb57ad776

GitOrigin-RevId: ba883fbf348a3deee88639b9aed4ea9709aa25be