Fix AndroidModuleNode.contains() on nested relocated modules
and similar cases when a nested module includes files which are not
under the roots of the parent module.

This change adds a new test project `projects/linked` with the structure
described above. (Note, that this change removes a check that ensures
that a test project contains a build file in the root directory. This is
not longer applicable and an invalid Gradle project won't sync anyway.)

Bug: n/a
Test: AndroidProjectViewNodeConsistencyTest
Change-Id: Iaf5fe4ce1f9eb2e0dd2037ba832012b28c6aea90