Fix inconsistent behavior of project view nodes
Project view nodes should have consistent implementation of
- contains()
- canRepresent()
- getVirtualFile()
- and those of nodes returned by getChildren() (directly or indirectly)

The platform uses these methods to locate nodes by a file. A node
revealing some details of a file must be locatable by it in order to be
correctly refreshed when the virtual file system changes.

This change adds a test that verifies that nodes representing a
sample project can be located by files they expose via:
- ProjectViewNode.getVirtualFile()  // the platform
- FolderGroupNode.getFolders()  // Android Studio
- FileGroupNode.getFiles()  // Android Studio

Bug: n/a
Test: AndroidProjectViewNodeConsistencyTest
Change-Id: I87ad4051909df453c0bf0e182077d23cbe9e2e44