Removing the ability to rename resources from Xml tag names
This feature clashes with a update to the platform that causes a
bug when more than 1 rename handler is available at one time.
XmlTagRename handler will be the only handler available, and users
will have to place the caret in the "name" Xml attribute value to
rename the resource.

Find usages remains the same, as the underlying platform issue does
not affect this action.

For 4.1 onwards, the change had to be via the RESOLVE_USING_REPOS
unified resources pipeline.

Bug: 153850296
Test: Regression test and disabled tests in AndroidRenameTest.java
Change-Id: Ia7409b18c42382a5b34c2cf272de67d29ba4970c