Implement retention action in a separate module
Android Test Retention is the project to take snapshots on test failures
and load it back for debug later. This commit implements the action to
push the test failure snapshot and load it in the emulator.

The current version still requires the user to configure a compatible
emulator before using Android Test Retention. The next step is to
automatically configure the emulator for the user, as well as
automatically connect debuggers after snapshot loads.

Bug: 152773288
Test: FindEmulatorAndSetupRetentionTest
Change-Id: I46299a849ac8d4deacd04d81569c21c9ecb448f1