Don't drop messages for other windows
If we get updates for more than one window at once (e.g. if we
connect to an app with a dialog open, or if a dialog is opened and
at the same time the original window is updated), previously we might
end up only processing the update for one (the latest received),
resulting in the dialog or in some cases the original window not being
shown.
Keeping a map of window ids to update times instead of just a single
time.

Test: Added a test
Bug: 152802794
Change-Id: I26da094e73c7773ebb669e8e510dada0b5394360