[AUA] working undo for the refactoring operation
projectBuildModel.applyChanges() is a PsiSpoiling operation.  As far
as I can tell, even things like replacing a child node with another
node (maybe of a different type?) is PsiSpoiling -- so it's not
totally surprising to me that most of our manipulations in the Dsl
writer are considered to be spoiling.

GradleWrapper.saveProperties() is not just PsiSpoiling, but its
low-level operation on a java.io.File appears to destroy the system's
understanding of modified files completely.  Rewrite the
performAgpUpgrade method to operate on the PsiElement -- that we have
had to find in any case in order to display it in the UsageView --
which is not a PsiSpoiling operation.

Bug: 149565531
Test: manual
Change-Id: I34ee0e8832ef007c5c7752c25c528b9ff7904ab1

GitOrigin-RevId: 135bcb1661bca30630b8a1158f004a3241f6fb35