Add single/double click navigation
This CL adds the distinction of double click/single click navigation to
the Compose Preview. When using single click, the preview will only
navigate within the current file. On double click, the preview can
navigate to other files if the source is available.
This is similar to the interaction with <include> in the Layout Editor
surface and hopefully will avoid users being confused avoid the editor
navigating outside of the file all the time.

Bug: 156747549
Test: Added tests but marked as @Ignore as they will only work in dev12
Change-Id: I937ce6f5a4a09ba85d403822f5a315ad491cab76