Work around tool window getting created on non-EDT
As of IJ 201.6487.11, tool windows for the default Executors
are lazily created. This causes our execution of the attach
debugger run settings to create a tool window on a non-EDT,
thus throwing an exception.

We fix this by pre-creating the window in case it's not
created already (or gets the existing window if it is) before
executing the debugger attachment.

Fixes: 156727360
Test: on fresh studio, open app on device, and attach debugger
Change-Id: I567575dcd1ca3d4f300b023b8a0e5923c19069d1
(cherry picked from commit 0a45bde05c1c3d12db95644677c083eda192a22a)