Clean up LayoutElement subclasses; add a FileDescriptor for <merge> tag.
Now only tags for which it make sense(need to support layout attributes or can have views as subtags) extends LayoutElement.
Register MergeDomFileDescription for the case <merge> tag is the root tag in a layout file.
Make AndroidLayoutDomExtender register view-subtags for all LayoutElements that supports it.

All above is internal cleanup that helps enable AndroidLayoutDomExtender.

Bug: b/157044963
Test: existing, no user-visible changes
Change-Id: I1d78dd104bab1e051270d651995495c4851aa4e2

GitOrigin-RevId: 7acf4bb9bd121f1e8ee915265ea6edb8b99cebff