Add fallback icon for Drawable gutter and support theme attributes
Added a fallback icon for Drawables that fail to render. This is to
still allow to have the Drawable picker from the Gutter.

Added support for theme attributes (drawables only).

These changes should make it so the gutter icon action is always
available even when it fails to produce a preview for the Drawable
resource.

Bug: 156769367
Test: AndroidGutterIconAnnotatorTest.java
Change-Id: Ifa4c70a66f156b0548bb4e1aa53b74c1ae2d0bda