Discovery to ignore a stream's past events when it reconnects
This CL fixes an issue that only happens when the only Studio project is closed and then re-opened. In this scenario, AppInspectionDiscovery first learns the devices are disconnected when the project is closed, and when the Studio project is opened again, it learns the devices are connected (again). This violated the original assumption that the pipeline/daemon would maintain connection between Studio and the devices as long as the Studio application is alive. The reality is the connections are severed as soon as the last Studio project window is closed.

The fix is in AppInspectionDiscovery. It now is aware of the last_active_event_timestamp for each stream/device, so when a device re-connects, it uses that timestamp to filter out and ignore events that it had seen already.

Test: new AppInspectionDiscoveryHost test.
Fixes: 154092735
Change-Id: I3eb6c65134920457d0d56ed46f041309ac6d6e56