Fix navigation when duplicate files are found
When clicking on androidx.ui.foundation.Text we could end up navigating
to a Text.kt within the current project. This is to the filename index
lookup.

This CL refactors the code to make sure we are navigating to the same
file and also reduces the stop of the index search to just the current
module with dependencies.

Fixes: 157129712
Test: New regression test
Change-Id: I6bde2d1ee40f3860e99b7b8a12bf0c8ecd2bf0f4

GitOrigin-RevId: f8db9d1edd0a59ddfbee3f761755a42218ff6100