Avoid overly stretching number columns on wide screens
The stretching of columns when the window is resized seem to be
hinted by their relative max widths. The default max width
is ~2 billion. This CL allows specifying `maxWidth` in
`ColumnTreeBuilder`, and sets moderate max widths for number columns
in the memory profiler. This saves all the space for the remaining
column(s) while the user can manually stretch the number columns
up to some reasonable width.

This CL also adjusts the preferred width to avoid "..." in column
headers when there's enough space.

Bug: 154271035
Test: manual
Change-Id: I5dc63d8f79aa9dd8c9351020399b5f6278f305e8

GitOrigin-RevId: 54973290848e8878f8927c56a1b0224080d11042