Update Studio now abi filters are not nullable
In I6de394be8335a672243504190068cb064a46629d abi filters in the Android
Gradle Plugin are made not null, along with some tools that are shared
with Studio.

For older AGPs the Ide* models now normalize null to empty set, so the
usages no longer have to deal with null.

Test: Manually verified deploying using AGP 3.6.3.
Bug: 157151941
Change-Id: Iaab4da2be03a80a01103d8bc65cef976267aec71

GitOrigin-RevId: 697b2a338279b7154466af32a61c248f9e809541