Add a proper support for <view/> tag.
We had a dedicated DomElement for <view> tag, but didn't use it consistently.

This CL adds ViewTagDomFileDescription that parse top level <view> tag as a View element, not a default LayoutElement.
It also changes ViewClassConverter to support Views that are not required package name.

Mostly it's internal clean up, there are no user visible changes but a fixed bug.

Fixes: b/157232598
Test: AndroidLayoutDomTest.testViewHighlighting
Change-Id: I1f05bb4a09c671cfa41edf3eaa92325e988d005e

GitOrigin-RevId: 42346e894e13f0cf3e63f8a9e3e41039bb72f2dc