Introduce AndroidXmlLayoutViewDescriptor.
Before that change for view tag that we can't resolve to class we had AndroidAnyTagDescriptor.
We also use AndroidAnyTagDescriptor for all tags that we can't resolve(that don't have descriptors) in XML, not only related to layout.

If we use AndroidLayoutDomExtender there is always a valid CustomDomChildrenDescription for subtags.
Now for unresolved view we AndroidXmlLayoutViewDescriptor where .getViewClass() == null instead of AndroidAnyTagDescriptor.

It's internal clean up, there are no user visible changes.

Bug: b/157044963
Test: existing
Change-Id: Iddab5b4658e11fd5617c53b61eee02ae780da79c