[BA] Reuse Task details panel for task warnings
Use same detail panel to present task warnings as we use
for tasks on the Tasks page.
Basically that means that we will not show special panel for the warning
but the general task panel with warnings inlined. Note that task can
have several warnings and they all will be listed on details panel.
Since we name the task warning node for the task path it does make
more sense to present details information the same way as in Tasks view.
For us it will reduce code duplication and ease the maintenance.

Bug: 155802617
Test: visually verified
Change-Id: I03018ec958bc5a80eb1ea7736bc9392683c34a50

GitOrigin-RevId: ebe36b45872230d0badbdb6d403c2d7c9e30a88b