Improvements to running render actions
This CL refactors how async render actions are executed. Before this CL,
there was no timeout for actions waiting in the queue so, if one tasks
blocks the render thread forever, the tasks would keep piling up for
ever. We address that in two ways:
- We limit the maximum size of the queue to 50 pending tasks
(configurable).
- We now timeout tasks waiting in the queue so they will not wait there
forever anymore. We do this by introducing two timeouts, the queueing
timeout and the action timeout.

The queueing timeout is the maximum time we are willing for a certain
task to be waiting to be executed. If the time passes, the action is
removed from the queue and never executed.
The action timeout is the maximum time we are willing for the action to
execute its code. Currently we do not have the ability to interrupt
them, but if we do at some point, this will cancel the action.
For now, the action will run anyway but the timeout will be notified to
those listening to the CompletableFuture.

As part of this change we also tweak a bit the handling of errors thrown
by render. We now cache the render size when possible so the previews do
not randomly resize when we miss a frame (because of a timeout or
otherwise).

We also test the executor better and remove code duplication between the
RenderService, RenderTask and RenderExecutor.

Fixes: 157490719
Test: Added test
Change-Id: I04e2386fb0adce90dfdf3a9462f28cc2d19adaad

GitOrigin-RevId: c3a7d40c202d0701776a2e84aa5d51d3173f3f95