Improvements to running render actions
This CL refactors how async render actions are executed. Before this CL,
there was no timeout for actions waiting in the queue so, if one tasks
blocks the render thread forever, the tasks would keep piling up for
ever. We address that in two ways:
- We limit the maximum size of the queue to 50 pending tasks
(configurable).
- We now timeout tasks waiting in the queue so they will not wait there
forever anymore.

We also test the executor better and remove code duplication between the
RenderService, RenderTask and RenderExecutor.

Fixes: 157490719
Test: Added test
Change-Id: I10182b6425c92435b938c985983756176e666677

GitOrigin-RevId: e6cc65a8681139c6b71320de29c02adbdf81e30e