Throw exception when launching an inspector on a process that no longer exists and catch it.
This CL fixes an exception that is logged when Studio tries to launch an inspector on a process that no longer exists. It happens due to a race between the UI and discovery. For example: selecting a process on device X when X is shutting down.

Instead of logging it, this CL creates a new exception for it and catches it in AppInspectionView, where it will be ignored.

Test: N/A - no behavior change.
Fixes: 157512972
Change-Id: I116dad5964b1b738d40ec0b6b566d855a23d22a3
(cherry picked from commit 3cce45f005a54a3e34441f265666072cd71f8aa1)