[AUA] expire Project Upgrade notification
If we show the project upgrade notification, and the user then changes
their gradle files to update the AGP version manually, the resulting
dialog will be nonsensical, so expire on changes to any Gradle
files.  (This is arguably over-strong, but the alternative is too
weak).

Expire the upgrade notification on sync also.  If a user is re-syncing
because something else (e.g. version control) has changed the files,
and this evades our editing detection, expiring the notification on
sync will cause the notification to appear at the end of sync only if
it is still needed.

Bug: 157676004
Test: manual
Change-Id: I3306da3e2b6ab191b18be1a0f63b5bbb4c24b2f5
(cherry picked from commit 1e5a6e108757ae6a4c65aac143541b4fa7f81917)