[AUA] parameterize gradle version update
From the (future) assistant itself, it might well end up being
hard-coded to the latest Gradle version, and simply toggled; however,
the AndroidPluginVersionUpdater is a little more flexible in principle
even if in practice it is always called with the latest version or
null.  Allow callers to specify a gradle version, and propagate it
through to the usage info.

Bug: 149565531
Test: manual
Change-Id: I4d280274790fc6c575ab0fcb4915d8b272e5ca07