Improved detection of build for Compose Preview
The preview would usually tried to detect if the project has been built
via detecting the ClassNotFound exceptions. Since there are a number of
causes for those exceptions, the system did not work on all cases. For
example, if a class throws during construction, it would be interpreted
by the preview as the project not being built.

This CL changes removes the heuristic and checks the build state. If
there is no build state, it tries to find one of .class files to verify
if a compilation has happened.

Test: New tests
Bug: 157903701
Bug: 154609017
Change-Id: I5562b19ff570a5e8cc56d208985f78d1861c91f4

GitOrigin-RevId: ff6adf7948f2d50c261007925f83a1dadd2fc0c7