[TPService] Get port from TPD stdout
Instead of using a hardcoded port, now TPD will bind to an available one
and use the stdout to announce it. We then read the process stdout in
order to get the port to create the channel later.

Tested by manually running the IDE and killing the process during a
parsing and after it finished but before parsing a 2nd trace. Verified
that both times (using logs) that the daemon was bound to different
ports and we created the channels correctly.

Bug: 149379691
Test: manual (see above)
Change-Id: I852ade5bedc4de1e5a7f1bb4553c2ea684daef4a

GitOrigin-RevId: c5db3c27903f5128831ac498841e823d0967d93c