Deprecate TransactionGuard.submitTransaction usage
All the changes here are related to the deprecation of the
TransactionGuard.submitTransaction() methods related to this cherry
picked commit from JetBrains: http://go/androidcommit/1aabdc3 as part of
the 2020.1 merge.

For detailed information on how to adapt to the deprecated API, and an
explanation of the IDEA threading model (write-safe contexts, locks,
etc) please read this document: http://go/idea-threading-model

In some cases, one of the consequences of using invokeLater() instead
of submitTransaction() is that in tests, you might have to call
PlatformTestUtil.dispatchAllInvocationEventsInIdeEventQueue() to ensure
that the Runnables that are queued for execution on the EDT are
actually executed (since the tests run in a headless environment).
You can also call UIUtil.dispatchAllInvocationEvents() instead of
PlatformTestUtil.dispatchAllInvocationEventsInIdeEventQueue().
When using submitTransaction() in the past, this was not necessary.

The following tests, and test infrastructure are changed:

1. A corresponding change to the test ImportApkProjectTest.java must be
   made, which is in a different change ag/11416689 with the same topic
   "replace-some-deprecated-xacts".
2. AndroidGradleTestUtils.kt

The following tests pass:

- BlazeAndroidTestRunConfigurationStateTest.java

Bug: 156102335

Change-Id: I229e70601915eadb063c7169b7902278adca54d9