[Dsl] override fake names given a real name element
If we look up a properties element with a real GradleNameElement
argument, we are usually in the circumstance where we are creating
that property element.  However, under some circumstances that element
can already exist, but with a name not associated with a file
presence.  In those circumstances, update the name element with the
real GradleNameElement's details.

Bug: 155497681
Test: existing (new tests forthcoming)
Change-Id: I76123f07426a0df1aa8ce55d375bb6b55a783e29

GitOrigin-RevId: 086078c1ec57b782f8eed6af43588641c4e66053