Show "Stop interactive" always when in interactive mode
This CL makes the isInteractive part of the ComposePreviewManager state
so the action can look into the actual state of the surface and decide
when to show the button.
Since the interactive preview can take a few seconds to start, this CL
repurposes the InteractionMode enum to make a small state machine. The
button is only visible when the interactive mode is fully ready and not
while it's starting.
Refreshing keeps the button being displayed.

Fixes: 158040209
Test: Manually tested
Change-Id: I1e0f06cac5588eabbf7a46f48bc5583186d7b981