Show loading overlay when heap dump starts
This is a partial fix to the problem where it's not obvious
to the user what to do or expect after pressing the heap dump button,
because we're not jumping into the heap dump stage with the loading screen
right away.

Jumping into the heap dump stage right away turns out not trivial, so
this CL is a compromise that shows a "taking heap dump" overlay
while the heap dump is being taken.

Bug: 154270276
Test: manual
Change-Id: Iac6c9792f3d6601e39f47fbdd0ddfa8e0316a5ab

GitOrigin-RevId: 793398abc31262d045718b1a25f4ab7227b04f8b