Differentiate pragma update and pragma query
Pragma statements can be used to query data from the db or to update
variables. This CL differentiate between update pragma statements
and query pragma statements, so that we know if we need to
display results or not.

Test: UtilsTest.kt
Bug: 158227306
Change-Id: Ia1d90a2a494f02a389c37bf7847912e24500b1a3
(cherry picked from commit c342f1f07190ddbcebaae3796c5deb3ee159815f)