[BA] Show empty state for broken build
With this change we start updating the state on broken build as well.
In case of a broken build a previous report is replaced with a fake
'broken build' report. Corresponding message is displayed instead of
the Build Analyzer report.
If Build Analyzer tab was opened on broken build event then the report
is replaced with message about broken build.
If the tab was closed then no new tab is opened on broken build
and user has no way to open it with current implementation.
These report state logic changes are mirrored in
BuildAttributionStateReporter, hence WNA button states, making
it so that after 'report ready' state it can return to 'build broken'
state. Previously 'report ready' state was in fact terminal.

Bug: 158416657
Test: updated
Change-Id: I87346b0279aac8a173afa23eeeb2defa999b6678

GitOrigin-RevId: 0564593a228737ad7f79126c10daa6b97eb32d08