New LiveLiteralsTransform
Adds a new LiveLiteralsTransform class transform that rewrites constant
accesses into a global constant registry lookup. This allows dynamically
modifying constants for classes transformed with it.

The transformations are applied to all primitive types and to String.

This CL only includes tests for the transformation and does not include
it into the ModuleClassLoader yet.

Bug: 139475541
Test: Class transformation tests pass
Change-Id: I63ba73998db64c42e9cd4c9e19aa2592f4590f4b

GitOrigin-RevId: 90d24497d005d7d84186f2f2f15f15e381d663e2