Show quickfix to use project jdk for gradle jvm
... when an "Invalid Gradle JDK configuration" error is presented.

When an invalid Gradle JDK is set, we display this error message with a
link to the Gradle Settings tab in the Settings menu. After a
simplification this tab, this setting is no longer visible and thus
users can not change it unless they manually modify the .idea/gradle.xml
file.

This cl adds GradleJvmNotificationExtension to append a quick fix in the
notification that sets the value of gradleJvm to
ExternalSystemJdkUtil.USE_PROJECT_JDK, causing the project to use the
value from the PSD.

Bug: 156112443
Test: GradleJvmNotificationExtensionTest and UseProjectJdkAsGradleJvmListenerTest
Change-Id: I5eb7e42374559ee1ec3aa827234e7bc1a1035561