[AUA] add isApplicable function to component processors
Components (sub-RefactoringProcessors) of the AGP Upgrade Refactoring
Processor declare their applicability to a particular AGP
upgrade (from current to new) by implementing this boolean function.

Current components:

- classpath dependency: always true;
- GMaven buildscript->repositories declaration: current < 3.0.0;
- Gradle wrapper version: always true;
- Java8: if the upgrade is from 4.1.x or earlier to 4.2.0 or later;
- obsolete compile/runtime configurations: from earlier than 5.0.0 to
  later than 3.5.0 (deprecation period).

Set an isEnabled boolean flag in each component, lazily initialized
from the result of the isApplicable function.  This isEnabled flag is
currently unused, but the intent is that it will determine whether the
overall processor runs the component, and can be overridden by that
processor settings or other user interface.

Bug: 149565531
Test: manual
Change-Id: If36fdc0ff9179f9a5fbeb03b50fd67ed683b6df8

GitOrigin-RevId: ae8a4cbe658a58ab73ed302c440b598096438a11