Delay build status check until Smart
Checking the build status would fail if no build information is
available when running in dumb mode.
If the build check runs in dumb mode, we are not able to find the .class
files even when they are present in disk. This delays the check until
the dumb mode has ended.

Fixes: 158653269
Test: Manually checked
Change-Id: Ia68589225b887fc0903546eee282646d68d43296

GitOrigin-RevId: 01a95142aeab04c4fa9e4c8fd4c82085c7962507