Wait for the ddm lib client ready
Previously we set up a callback to get client ready signal in retention.
But it does not seem to catch all the new client signals we care about.
Thus we switch to a spin and wait approach.

Plus, even if we use a callback, there will be a spin and wait in
awaitWithCheckCanceled anyways.

Bug: 152773288
Test: FindEmulatorAndSetupRetention
Change-Id: I2e2108e090ce425b6ac258bb9a7e463a9b178d68

GitOrigin-RevId: 453b472730948330d903eedfdd0794d15988c3d0