[AUA] respect isEnabled per-component flag
Have the main AgpUpgradeRefactoringProcessor store individual
component processors.  The ClasspathDependency refactoring processor
is treated specially because it in some sense represents the overall
refactoring, and may be a `target' for it.  Other processors are
stored in a list, and iterated over.

This means that the component processors themselves are instantiated
once per overall refactoring, rather than once per find usages
operation, and consequently can be used to store and query state (such
as settings), including the isEnabled flag.  Centralize this logic in
a final implementation of findUsages(), which delegates to a new
findComponentUsages() function for overriding in individual component
processors.

Bug: 149565531
Test: manual
Change-Id: If0ffc51c3e2845a30aff4c43ac2f0875c131b960

GitOrigin-RevId: 84a668967ff2e4da03a78522969cbff4a221f7b4