[Gradle Sync] Cache build output-metadata parse result
This change implements basic caching of the Application ID in ensure
that we don't parse the build output listing file each time it is
requested.

The cache is implemented as a trainsiant field within each modules
AndroidModuleModel which when requested from the model either returns
the cached value or computes it by parsing the build output listing
file.

In order to invalidate this cache we create a Project service that
stores the timestamp of the last sync or build and if the value in the
cache came from a time before the last sync or build then it is
re-computed.

Bug: 159341725
Test: added unit test
Change-Id: I780fd1b0e2588652e2f0ae08a6b9f8ddabf3e3ed

GitOrigin-RevId: 8fd438d84430cac1a93c2ce05652bcb1495f46d6