[AUA] show a dialog box before refactoring
The list of actions shows through its selections those which are
applicable to the current upgrade.  Users might change those
selections, but the resulting refactoring action does not change as a
result (in this CL; a forthcoming one will enable this functionality).

The Cancel action exits the Upgrade process complete; the OK action
proceeds to the refactoring preview.

I don't anticipate this to be the final UI, but having something that
allows brave users to test the functionality would be good.

Bug: 149565531
Test: manual
Change-Id: I0308c32e9b185b702d11a0ed8510daca2eec21e7

GitOrigin-RevId: 51579128effbf850e2b34600d753307279795711