[AUA] add a Preview button
Acts the same as the OK button, but sets the preview flag of the
RefactoringProcessor before continuing.  In order for this to have any
effect, do not unconditionally set the preview flag in the
GradlePluginUpgrade flow.

Also rename the OK button to "Upgrade" in the UI to accord with the
other verbs on buttons in the dialog.

Bug: 149565531
Test: manual
Change-Id: Iea0952769a4edabba6016c9ed3eb0d86dbb4be56