[AUA] parameterise GMaven refactoring by gradleVersion
Make a gradle version (though not yet, sadly, *the* gradle version) a
parameter of the GMaven refactoring.  The refactoring theoretically
needs this because the syntax for a GMaven declaration in the
buildscript repositories block changed in Gradle version 4.0.  Since
these refactorings are unlikely to run in a version of Android Studio
lower than 4.2 (not supporting versions of Gradle below 6.5 or so) the
issue is theoretical and we can get away with using the minimum
supported version for this.  (Comments explain why using the project's
current or to-be-upgraded versions of Gradle are not quite
straightforward).

Bug: 149565531
Test: manual
Change-Id: I1b6550381411462fc5e4abf4e296ab5c9d963d9c