[TP] Split process metadata fetching to its own method
As we want for the loadTrace to return statistics about the parsing of
the loaded trace (or errors) we need to move the process metadata
fetching to its own method.

This is something we wanted to do anyway, since not all users of TPD
actually need the list of the process, for example
NativeAllocationSampleCaptureObject.

Bug: 156578844
Test: existing
Change-Id: I2dbe0f5b1077d8c623b0a3df5e47dae1836af839

GitOrigin-RevId: 57f87336f9f1e3a850b10024667aed97f2c84658