Intercept Compose Animation (un)subscribe calls
Create PreviewAnimationClockMethodTransform to intercept calls to
PreviewAnimationClock (from Compose ui-tooling) subscription methods using
bytecode manipulation. These calls are intercepted so we can notify the
ComposePreviewAnimationManager that an animation has subscribed and need to
be displayed in the animation inspector. Similarly, the manager will be
notified if an animation unsubscribes to the clock, so it can close the
corresponding tab in the inspector.

Test: ComposePreviewAnimationManagerTest
Bug: 158752769
Change-Id: I23a85015ae0e113315cfeb35c0fed448c2ac44d1

GitOrigin-RevId: 58f0b5a1707e5606b01e50f1574d2261c592da6d