Fix needless duplications of `ClassEntry` instances
The original code created and returned fresh instances of `ClassEntry`
every time the class was registered, which were saved in
`HeapDumpInstanceObject` instances. As a consequence, the number of
instances of `ClassEntry` was roughly the same as
`HeapDumpInstanceObject`, even though it should only be the number of
classes used by the program.

Impact of this CL on loading a ~21MB `.hprof` file:
* ClassEntry:
  - old: ~199K instances, ~28MB
  - new: ~10K instances, 2MB
* HeapDumpInstanceObject:
  - old: ~201K instances, 35M
  - new: ~201K instances, 9M

With this CL, objects in the profiler package are no longer the
main culprits in memory hog. `HeapDumpInstanceObject` is now
a fairly thin wrap around `perflib.heap.Instance`.

Bug: 152344964
Bug: 156553455
Test: existing, manual inspection of memory consumption
Change-Id: I2e1ea7f9d3033ca891941d59f06b5dac7a83fce7
(cherry picked from commit 093a21459bdbb97435172d33c04802ee76fb6bf1)